refactor(cli): split elicitation-repl.ts into mode modules (#367)
